+++ GrafoExplorer/ed/PriorityQueue.h
@@ -7,7 +7,7 @@ public:
 	PriorityQueue(const PriorityQueue<E>&) = delete;
 	void operator=(const PriorityQueue<E>&) = delete;
 	virtual ~PriorityQueue() = default;
-	virtual void insert(E element, int priority) = 0;
+	virtual void insert(E element, float priority) = 0;
 	virtual E min() = 0;
 	virtual E removeMin() = 0;
 	virtual void clear() = 0;
